Move file fixtures to dedicated module
